app/src/main/java/com/majazeh/risloo/Utils/Managers/VersionManager.java
package com.majazeh.risloo.utils.managers;

import android.content.Context;
import android.content.pm.PackageInfo;
import android.content.pm.PackageManager;

import com.majazeh.risloo.R;

public class VersionManager {

    /*
    ---------- Func's ----------
    */

    public static int versionCode(Context context) {
        try {
            PackageInfo packageInfo = context.getPackageManager().getPackageInfo(context.getPackageName(), 0);

            return packageInfo.versionCode;
        } catch (PackageManager.NameNotFoundException e) {
            e.printStackTrace();
        }

        return -1;
    }

    public static String versionName(Context context) {
        try {
            PackageInfo packageInfo = context.getPackageManager().getPackageInfo(context.getPackageName(), 0);

            return packageInfo.versionName;
        } catch (PackageManager.NameNotFoundException e) {
            e.printStackTrace();
        }

        return "";
    }

    public static String versionNamePrefix(Context context) {
        try {
            PackageInfo packageInfo = context.getPackageManager().getPackageInfo(context.getPackageName(), 0);

            if (packageInfo.versionName.contains("-"))
                return StringManager.prefix(packageInfo.versionName, '-');
            else
                return packageInfo.versionName;

        } catch (PackageManager.NameNotFoundException e) {
            e.printStackTrace();
        }

        return "";
    }

    public static String versionNameDesc(Context context) {
        try {
            PackageInfo packageInfo = context.getPackageManager().getPackageInfo(context.getPackageName(), 0);

            if (packageInfo.versionName.contains("-"))
                return context.getResources().getString(R.string.AppVersion) + " " + StringManager.prefix(packageInfo.versionName, '-');
            else
                return context.getResources().getString(R.string.AppVersion) + " " + packageInfo.versionName;

        } catch (PackageManager.NameNotFoundException e) {
            e.printStackTrace();
        }

        return "";
    }

}